Add javadoc to the AuthorizationEngine interface
This commit adds javadocs to the AuthorizationEngine interface aimed at
developers of an authorization engine. Additionally, some classes were
also moved to the core project so that they are ready to be exposed
once we allow authorization engines to be plugged in.
